refactor scripts and browser rail
