+++ scripts/scripts.js
@@ -505,70 +505,6 @@ export function htmlToElement(html) {
   return template.content.firstElementChild;
 }
 
-export async function loadPrevNextBtn() {
-  let placeholders = {};
-  try {
-    // eslint-disable-next-line no-use-before-define
-    placeholders = await fetchLanguagePlaceholders();
-  } catch (err) {
-    // eslint-disable-next-line no-console
-    console.error('Error fetching placeholders:', err);
-  }
-  const mainDoc = document.querySelector('main > div.content-section-last');
-  if (!mainDoc) return;
-
-  const prevPageMeta = document.querySelector('meta[name="prev-page"]');
-  const nextPageMeta = document.querySelector('meta[name="next-page"]');
-  const prevPageMetaContent = prevPageMeta?.getAttribute('content').trim().split('.html')[0];
-  const nextPageMetaContent = nextPageMeta?.getAttribute('content').trim().split('.html')[0];
-  const PREV_PAGE = placeholders?.previousPage;
-  const NEXT_PAGE = placeholders?.nextPage;
-
-  if (prevPageMeta || nextPageMeta) {
-    if (prevPageMetaContent === '' && nextPageMetaContent === '') return;
-
-    const docPagination = createTag('div', { class: 'doc-pagination' });
-    const btnGotoLeft = createTag('div', { class: 'btn-goto is-left-desktop' });
-
-    const anchorLeftAttr = {
-      // eslint-disable-next-line no-use-before-define
-      href: `${rewriteDocsPath(prevPageMetaContent)}`,
-      class: 'pagination-btn',
-    };
-    const anchorLeft = createTag('a', anchorLeftAttr);
-    const spanLeft = createTag('span', '', PREV_PAGE);
-
-    anchorLeft.append(spanLeft);
-    btnGotoLeft.append(anchorLeft);
-
-    const btnGotoRight = createTag('div', {
-      class: 'btn-goto is-right-desktop',
-    });
-
-    const anchorRightAttr = {
-      // eslint-disable-next-line no-use-before-define
-      href: `${rewriteDocsPath(nextPageMetaContent)}`,
-      class: 'pagination-btn',
-    };
-    const anchorRight = createTag('a', anchorRightAttr);
-    const spanRight = createTag('span', '', NEXT_PAGE);
-
-    anchorRight.append(spanRight);
-    btnGotoRight.append(anchorRight);
-
-    if (!prevPageMeta || prevPageMetaContent === '') {
-      anchorLeft.classList.add('is-disabled');
-    }
-
-    if (!nextPageMeta || nextPageMetaContent === '') {
-      anchorRight.classList.add('is-disabled');
-    }
-
-    docPagination.append(btnGotoLeft, btnGotoRight);
-    mainDoc.append(docPagination);
-  }
-}
-
 /**
  * Copies all meta tags to window.EXL_META
  * These are consumed by Qualtrics to pass additional data along with the feedback survey.
@@ -624,20 +560,6 @@ function showBrowseBackgroundGraphic() {
   }
 }
 
-async function loadPage() {
-  await loadEager(document);
-  await loadLazy(document);
-  loadRails();
-  loadDelayed();
-  showBrowseBackgroundGraphic();
-  await loadPrevNextBtn();
-}
-
-// load the page unless DO_NOT_LOAD_PAGE is set - used for existing EXLM pages POC
-if (!window.hlx.DO_NOT_LOAD_PAGE) {
-  loadPage();
-}
-
 /**
  * Helper function that converts an AEM path into an EDS path.
  */
@@ -679,49 +601,6 @@ export function rewriteDocsPath(docsPath) {
   return url.toString().replace(PROD_BASE, ''); // always remove PROD_BASE if exists
 }
 
-/**
- * Helper function thats returns a list of all products
- * - below <lang>/browse/<product-page>
- * - To get added, the product page must be published
- * - Product pages listed in <lang>/browse/top-products are put at the the top
- *   in the order they appear in top-products
- * - the top product list can point to sub product pages
- */
-export async function getProducts() {
-  // get language
-  const { lang } = getPathDetails();
-  const ffetch = (await ffetchModulePromise).default;
-  // load the <lang>/top_product list
-  const topProducts = await ffetch(`/${lang}/top-products.json`).all();
-  // get all indexed pages below <lang>/browse
-  const publishedPages = await ffetch(`/${lang}/browse-index.json`).all();
-
-  // add all published top products to final list
-  const finalProducts = topProducts.filter((topProduct) => {
-    // check if top product is in published list
-    const found = publishedPages.find((elem) => elem.path === topProduct.path);
-    if (found) {
-      // keep original title if no nav title is set
-      if (!topProduct.title) topProduct.title = found.title;
-      // set marker for featured product
-      topProduct.featured = true;
-      // remove it from publishedProducts list
-      publishedPages.splice(publishedPages.indexOf(found), 1);
-      return true;
-    }
-    return false;
-  });
-
-  // for the rest only keep main product pages (<lang>/browse/<main-product-page>)
-  const publishedMainProducts = publishedPages
-    .filter((page) => page.path.split('/').length === 4)
-    // sort alphabetically
-    .sort((productA, productB) => productA.path.localeCompare(productB.path));
-  // append remaining published products to final list
-  finalProducts.push(...publishedMainProducts);
-  return finalProducts;
-}
-
 export async function fetchLanguagePlaceholders() {
   const { lang } = getPathDetails();
   try {
@@ -748,4 +627,28 @@ export async function getLanguageCode() {
   const langObj = langMap.find((item) => item.key === lang);
   const langCode = langObj ? langObj.value : lang;
   return langCode;
+}
+
+async function loadDefaultModule(jsPath) {
+  try {
+    const mod = await import(jsPath);
+    if (mod.default) await mod.default();
+  } catch (error) {
+    // eslint-disable-next-line no-console
+    console.log(`failed to load module for ${jsPath}`, error);
+  }
+}
+
+async function loadPage() {
+  await loadEager(document);
+  await loadLazy(document);
+  loadRails();
+  loadDelayed();
+  showBrowseBackgroundGraphic();
+  loadDefaultModule(`${window.hlx.codeBasePath}/scripts/prev-next-btn.js`);
+}
+
+// load the page unless DO_NOT_LOAD_PAGE is set - used for existing EXLM pages POC
+if (!window.hlx.DO_NOT_LOAD_PAGE) {
+  loadPage();
 }
